backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
alembic>=1.13,<2.0
pydantic-settings>=2.2,<3.0
python-multipart>=0.0.9,<1.0
redis>=5.0,<7.0
celery>=5.4,<6.0
aiogram>=3.7,<4.0
passlib>=1.7,<2.0
greenlet>=3.0,<4.0
